requirements/base.txt
django==4.2.11
psycopg2-binary==2.9.9
pillow==10.2.0
redis==5.0.2
python-environ==0.4.54
django_redis==5.4.0
django-filter==23.5
django-modeltranslation==0.18.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-recaptcha==4.0.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.36
django-resized==1.0.2
djangorestframework-simplejwt==5.3.1
django-image-uploader-widget==0.5.0
django-unfold==0.22.0
